+++ programs/drift/src/controller/orders.rs
@@ -296,6 +296,30 @@ pub fn place_perp_order(
         return Err(ErrorCode::InvalidOrderNotRiskReducing);
     }
 
+    let max_oi = market.amm.max_open_interest;
+    if max_oi != 0 && !risk_decreasing {
+        let oi_plus_order = match params.direction {
+            PositionDirection::Long => market
+                .amm
+                .base_asset_amount_long
+                .safe_add(order_base_asset_amount.cast()?)?
+                .unsigned_abs(),
+            PositionDirection::Short => market
+                .amm
+                .base_asset_amount_short
+                .safe_sub(order_base_asset_amount.cast()?)?
+                .unsigned_abs(),
+        };
+
+        validate!(
+            oi_plus_order <= max_oi,
+            ErrorCode::MaxOpenInterest,
+            "Order Base Amount={} could breach Max Open Interest for Perp Market={}",
+            order_base_asset_amount,
+            params.market_index
+        )?;
+    }
+
     let (taker, taker_order, maker, maker_order) =
         get_taker_and_maker_for_order_record(&user_key, &new_order);